Initial resolveAvailability implementation
Given a SchedulingParameters with recurring availability windows, and a
time range to search, find the intervals within the range corresponding
to the available times of that SchedulingParameters.

This is a building block for the $find operation.

Timezone math is surprisingly difficult with the primitives currently
available in baseline JS! Being able to specify a wall clock time and a
timezone to apply it in while accounting for weirdness like DST is quite
irksome.

While there are libraries like `date-fns` and `luxor` that provide
implementations of this around native `Date` instances, the direction
that the broader ecosystem is moving in is towards new "Temporal" types.
This is currently a stage-3 proposal that is live in Firefox, is built
in to V8 (unflagged in
https://issues.chromium.org/issues/401065166#comment106), and so seems
likely to be included as early as Node 26.

Introducing `temporal-polyfill` to our server dependencies lets us start
using the API today, and should be easy to remove once all our supported
node versions include the API natively.

Signed-off-by: Noah Silas <noah@medplum.com>