Schedule $find handler initial implementation
This is implemented as a stepping stone to future Appointment/Slot $find
operations. Working on those will entail computing availability for
multiple schedules and intersecting them; this endpoint that only
considers a single schedule is necessarily simpler.

This is still incomplete - it only handles the "default" availability
type, not anything scoped by "service-type", and so doesn't respect
ActivityDefinition resources that should affect availability or
alignment. It doesn't handle booking limits. It doesn't make a decision
on how to interpret timezone information when the schedule is missing
explicit timezone annotation. This is intended only as a checkpoint on
the way to a more full implementation coming soon.

Signed-off-by: Noah Silas <noah@medplum.com>