Display Timing.repeat.timeOfDay independently of timezone
A `timeOfDay` is represented with a string like "08:00", which should
represent "8 AM" without being tied to a specific timezone.

When the `formatTime` function was being used to represent this, it was
including the viewer's timezone offset in the calculations. In my case
(PST timezone, which is offset UTC-8) this would result in the displayed
time being offset by 8 hours.

I didn't want to introduce a potentially breaking change to the
`formatTime` function, so I've introduced a new `formatWallTime`
function that should stay timezone agnostic.

Signed-off-by: Noah Silas <noah@medplum.com>